requirements.txt
streamlit==1.44.1
tensorflow-cpu==2.16.1
numpy==1.26.4
Pillow==10.4.0
h5py==3.11.0
